chat/requirements.txt
absl-py==2.2.1
accelerate==1.9.0
aiocache==0.12.3
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.11
aiosignal==1.4.0
alembic==1.14.0
annotated-types==0.7.0
anthropic==0.59.0
anyio==4.9.0
appdirs==1.4.4
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.9.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.4.1
av==15.0.0
azure-ai-documentintelligence==1.0.2
azure-core==1.35.0
azure-identity==1.20.0
azure-storage-blob==12.24.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
bitarray==3.5.2
black==25.1.0
blinker==1.9.0
boto3==1.35.53
botocore==1.35.99
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.3.3
certifi==2025.7.9
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colbert-ai==0.2.21
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.5
ctranslate2==4.6.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.0.0
ddgs==9.0.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
Django==5.2.4
dnspython==2.7.0
docker==7.1.0
docx2txt==0.8
duckdb==0.8.1
durationpy==0.10
ecdsa==0.19.1
einops==0.8.1
elastic-transport==8.17.1
elasticsearch==9.0.1
emoji==2.14.1
et_xmlfile==2.0.0
Events==0.5
executing==2.2.0
fake-useragent==2.1.0
fastapi==0.115.7
faster-whisper==1.1.1
filelock==3.18.0
filetype==1.2.0
firecrawl-py==1.12.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
fpdf2==2.8.2
frozenlist==1.7.0
fs==2.4.16
fsspec==2025.3.0
ftfy==6.2.3
gast==0.6.0
gcp-storage-emulator==2024.8.3
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.15.0
google-generativeai==0.8.5
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.16.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.2
Jinja2==3.1.5
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==2.15.0
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.72
langchain-text-splitters==0.3.8
langdetect==1.0.9
langfuse==2.44.0
langsmith==0.4.8
ldap3==2.9.1
libclang==18.1.1
loguru==0.7.3
lxml==6.0.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
ml-dtypes==0.3.2
mmh3==5.1.0
moto==5.1.8
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
msvc-runtime==14.34.31931
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.5
ninja==1.11.1.4
nltk==3.9.1
numpy==1.26.4
nvidia-ml-py==12.535.161
nvitop==1.3.2
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.20.1
open-webui==0.6.18
openai==1.95.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opensearch-py==2.8.0
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-aiohttp-client==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-instrumentation-httpx==0.53b1
opentelemetry-instrumentation-logging==0.53b1
opentelemetry-instrumentation-redis==0.53b1
opentelemetry-instrumentation-requests==0.53b1
opentelemetry-instrumentation-sqlalchemy==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
opt_einsum==3.4.0
optree==0.14.1
orjson==3.11.0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
peewee==3.18.1
peewee-migrate==1.12.2
pgvector==0.4.0
pillow==11.3.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
playwright==1.49.1
pluggy==1.6.0
portalocker==2.10.1
posthog==5.4.0
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
py-partiql-parser==0.6.1
pyarrow==21.0.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pycrdt==0.12.25
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==12.0.0
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.14.2
pymilvus==2.5.0
pymongo==4.13.2
PyMySQL==1.1.1
pyodide-http==0.2.2
pypandoc==1.15
pyparsing==3.2.3
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySimpleGUI==5.0.6
pytest==8.3.5
pytest-docker==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-iso639==2025.2.18
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
python-socketio==5.13.0
pytube==15.0.0
pytz==2025.2
pywin32==310
pyxlsb==1.0.10
PyYAML==6.0.2
pyzmq==26.4.0
qdrant-client==1.14.3
rank-bm25==0.2.2
RapidFuzz==3.13.0
rapidocr-onnxruntime==1.4.4
razorpay==1.4.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.7
RestrictedPython==8.0
rich==13.9.4
rsa==4.9
s3transfer==0.10.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
shapely==2.1.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.7
SQLAlchemy==2.0.38
sqlite-bro==0.12.2
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.45.3
starlette-compress==1.6.0
sympy==1.14.0
tenacity==9.1.2
tencentcloud-sdk-python==3.0.1336
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
unstructured==0.16.17
unstructured-client==0.40.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.2
validators==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
windows-curses==2.3.3
winpython==6.4.20230625
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
xlsxwriter==3.2.5
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.20.1
youtube-transcript-api==1.1.0
zipp==3.23.0
zstandard==0.23.0
